+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Development version
+* `generic` previewer now understands `<meta property="article:published_time" …>`  OpenGraph tags 
+
 ## Version 1.0
 
 * First release